Leave the generated points of ImageData and RectilinearGrid to VTK
Reading `.points` to learn its dtype materializes the lazy structured array VTK
9.4+ returns, which OOM'd CI trying to allocate 25.8 GiB of meshgrid for a
voxelize reference volume. `'preserve'` now preserves only the dtype of points a
mesh actually stores, so a grid input is never read for one.

That also keeps `voxelize` single precision when its input is: it builds an
ImageData on the way, whose generated double would otherwise widen an unrelated
output. Image filters go back to matching main, and `cast_to_unstructured_grid`
keeps its own explicit double, where the coordinates are the data being cast.

Fix two docs-build failures alongside: `pyvista.PyVistaPrecisionWarning` had no
autoclass to resolve against, and `numpy.float32`/`float64` are `py:attribute`
in the intersphinx inventory rather than `py:class`.

+++ tests/core/test_rectilinear_grid_filters.py
@@ -84,5 +84,7 @@ def test_to_tetrahedra_pass_cell_data(tiny_rectilinear):
 
 
 def test_to_tetrahedra_points_dtype(rectilinear):
+    # RectilinearGrid generates its points rather than storing them, so 'preserve'
+    # constrains nothing here and vtkRectilinearGridToTetrahedra picks single
     assert rectilinear.points.dtype == np.double
-    assert rectilinear.to_tetrahedra().points.dtype == np.double
+    assert rectilinear.to_tetrahedra().points.dtype == np.single
